+++ src/app/components/file-info-modal.js
@@ -2,7 +2,7 @@
  * Modal with detailed views about file or directory.
  *
  * @author Jakub Liput
- * @copyright (C) 2019-2022 ACK CYFRONET AGH
+ * @copyright (C) 2019-2023 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
@@ -22,6 +22,8 @@ import {
   bool,
   equal,
   not,
+  array,
+  conditional,
 } from 'ember-awesome-macros';
 import EmberObject, { computed, get, set, getProperties, observer } from '@ember/object';
 import resolveFilePath, { stringifyFilePath } from 'oneprovider-gui/utils/resolve-file-path';
@@ -52,6 +54,17 @@ const mixins = [
  * @property {string} [tabClass]
  */
 
+/**
+ * @typedef {'show'|'download'} FileInfoModal.FileLinkType
+ */
+
+/**
+ * @typedef {Object} FileInfoModal.FileLinkOption
+ * @property {string} url
+ * @property {FileInfoModal.FileLinkType} type
+ * @property {SafeString} label
+ */
+
 export default Component.extend(...mixins, {
   i18n: service(),
   fileManager: service(),
@@ -59,6 +72,7 @@ export default Component.extend(...mixins, {
   spaceManager: service(),
   storageManager: service(),
   providerManager: service(),
+  appProxy: service(),
 
   open: false,
 
@@ -74,10 +88,9 @@ export default Component.extend(...mixins, {
   files: undefined,
 
   /**
-   * @virtual optional
-   * @type {Boolean}
+   * @type {Utils.FilesystemBrowserModel}
    */
-  previewMode: false,
+  browserModel: undefined,
 
   /**
    * Share to which opened file belongs to (should be provided in preview mode),
@@ -149,6 +162,11 @@ export default Component.extend(...mixins, {
   getProvidersUrl: notImplementedIgnore,
 
   /**
+   * @type {FileInfoModal.FileLinkType}
+   */
+  selectedFileLinkType: null,
+
+  /**
    * @type {FileInfoTabId}
    */
   activeTab: 'general',
@@ -163,6 +181,11 @@ export default Component.extend(...mixins, {
    * @type {Number}
    */
   hardlinksLimit: 100,
+
+  /**
+   * @type {ComputedProperty<boolean>}
+   */
+  previewMode: reads('browserModel.previewMode'),
 
   tabItemsIcons: computed(function tabItemsIcons() {
     const icons = {
@@ -204,6 +227,57 @@ export default Component.extend(...mixins, {
       defaultValue: this.t('fileType.file'),
     }));
   }),
+
+  /**
+   * @type {Array<FileInfoModal.FileLinkType>}
+   */
+  availableFileLinkTypes: Object.freeze(['show', 'download']),
+
+  /**
+   * @type {ComputedProperty<Array<FileInfoModal.FileLinkOption>>}
+   */
+  availableFileLinkOptions: computed(
+    'availableFileLinkTypes',
+    'itemType',
+    'previewMode',
+    function availableFileLinkOptions() {
+      if (
+        !this.file?.type ||
+        !this.availableFileLinkTypes ||
+        // TODO: VFS-11156 Implement shared files global URLs
+        this.previewMode
+      ) {
+        return [];
+      }
+      return this.availableFileLinkTypes.map(fileLinkType => ({
+        type: fileLinkType,
+        url: this.appProxy.callParent('getFileGoToUrl', {
+          fileId: this.file.cdmiObjectId,
+          fileAction: fileLinkType,
+        }),
+        label: this.t(`fileLinkLabel.${fileLinkType}`),
+        tip: this.t(`fileLinkTip.${fileLinkType}.${this.itemType}`, { defaultValue: '' }),
+      }));
+    }
+  ),
+
+  /**
+   * @type {ComputedProperty<FileInfoModal.FileLinkType>}
+   */
+  effSelectedFileLinkType: conditional(
+    array.includes('availableFileLinkTypes', 'selectedFileLinkType'),
+    'selectedFileLinkType',
+    'availableFileLinkTypes.firstObject'
+  ),
+
+  effSelectedFileLinkOption: computed(
+    'effSelectedFileLinkType',
+    function effSelectedFileLinkOption() {
+      return this.availableFileLinkOptions.find(option =>
+        option.type === this.effSelectedFileLinkType
+      );
+    }
+  ),
 
   headerText: computed('typeTranslation', function headerText() {
     if (this.typeTranslation) {
@@ -321,12 +395,11 @@ export default Component.extend(...mixins, {
           });
 
           const provider = await get(storage, 'provider');
-          const providerName = get(provider, 'name');
           const storageName = get(storage, 'name');
 
           const storageNameWithPath = {
             storageName,
-            providerName,
+            provider,
             path: locationsPerStorage[storageId],
           };
 
@@ -335,6 +408,16 @@ export default Component.extend(...mixins, {
           } else {
             locationsPerProviderWithStorageName[providerId] = [storageNameWithPath];
           }
+        }
+
+        if (locationsPerProvider[providerId].error) {
+          const provider = await this.providerManager.getProviderById(providerId);
+          const providerError = {
+            storageName: null,
+            provider,
+            error: locationsPerProvider[providerId].error,
+          };
+          locationsPerProviderWithStorageName[providerId] = [providerError];
         }
       }
 
@@ -358,22 +441,6 @@ export default Component.extend(...mixins, {
       return Object.keys(storageLocationsPerProvider).length;
     }
   ),
-
-  fileGuiUrlProxy: promise.object(computed('file.entityId', async function fileGuiUrl() {
-    const {
-      file,
-      getDataUrl,
-    } = this.getProperties('file', 'getDataUrl');
-    if (!file || !getDataUrl || getDataUrl === notImplementedThrow) {
-      return;
-    }
-    return await getDataUrl({
-      dir: null,
-      selected: [get(file, 'entityId')],
-    });
-  })),
-
-  fileGuiUrl: reads('fileGuiUrlProxy.content'),
 
   symlinkTargetPath: computed(
     'file.{type,targetPath}',
@@ -545,13 +612,14 @@ export default Component.extend(...mixins, {
    * @type {ComputedProperty<boolean>}
    */
   isSizeTabVisible: computed(
-    'previewMode',
+    'browserModel.isDirStatsFeatureHidden',
     'isMultiFile',
     'file.effFile.type',
     'itemType',
     function isSizeTabVisible() {
       const effItemType = get(this.file, 'effFile.type') || 'file';
-      return !this.previewMode && !this.isMultiFile && effItemType !== 'file' &&
+      return !this.browserModel?.isDirStatsFeatureHidden &&
+        !this.isMultiFile && effItemType !== 'file' &&
         this.itemType !== 'symlink';
     }
   ),
@@ -893,6 +961,13 @@ export default Component.extend(...mixins, {
     },
     getProvidersUrl(...args) {
       return this.get('getProvidersUrl')(...args);
+    },
+
+    /**
+     * @param {ComputedProperty<FileInfoModal.FileLinkType>} fileLinkType
+     */
+    changeSelectedFileLinkOption(fileLinkOption) {
+      this.set('selectedFileLinkType', fileLinkOption.type);
     },
   },
 });