+++ src/app/services/onedata-websocket-error-handler.js
@@ -1,33 +1,14 @@
 /**
- * Set global variables to inform user when WebSocket was exited abnormally
+ * Extends generic OnedataWebsocketErrorHandler with reconnector GUI support.
  *
  * @author Jakub Liput
- * @copyright (C) 2019 ACK CYFRONET AGH
+ * @copyright (C) 2023 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
 import OnedataWebsocketErrorHandler from 'onedata-gui-websocket-client/services/onedata-websocket-error-handler';
-import {
-  GOING_AWAY,
-} from 'onedata-gui-websocket-client/utils/websocket-close-event-codes';
+import ReconnectingWebsocketErrorHandlerMixin from 'onedata-gui-common/mixins/reconnecting-websocket-error-handler';
 
-export default OnedataWebsocketErrorHandler.extend({
-  /**
-   * @override
-   *
-   * @param {CloseEvent} closeEvent
-   * @param {boolean} openingCompleted
-   */
-  abnormalClose(closeEvent, openingCompleted) {
-    if (closeEvent && closeEvent.code === GOING_AWAY) {
-      console.debug(
-        'service:onedata-websocket-error-handler#abnormalClose: GOING_AWAY code, ignoring'
-      );
-    } else {
-      this.setProperties({
-        currentCloseEvent: closeEvent,
-        currentOpeningCompleted: openingCompleted,
-      });
-    }
-  },
-});
+export default OnedataWebsocketErrorHandler.extend(
+  ReconnectingWebsocketErrorHandlerMixin,
+);