+++ src/tests/integration/components/file-info-modal-test.js
@@ -17,6 +17,7 @@ import FileDistributionHelper from '../../helpers/file-distribution';
 import createSpace from '../../helpers/create-space';
 import DefaultUser from '../../helpers/default-user';
 import globals from 'onedata-gui-common/utils/globals';
+import FilesystemBrowserModel from 'oneprovider-gui/utils/filesystem-browser-model';
 
 const storageLocations = {
   locationsPerProvider: {
@@ -103,7 +104,11 @@ const file1 = {
 };
 
 describe('Integration | Component | file-info-modal', function () {
-  setupRenderingTest();
+  const { afterEach } = setupRenderingTest();
+
+  afterEach(function () {
+    this.browserModel?.destroy?.();
+  });
 
   // NOTE: context is not used for async render tests, because mocha's context is buggy
 
@@ -705,6 +710,7 @@ async function renderComponent() {
     getDataUrl=getDataUrl
     storageLocationsProxy=storageLocationsProxy
     tabOptions=tabOptions
+    browserModel=browserModel
   }}`);
 }
 
@@ -828,6 +834,11 @@ async function givenDefaultStubs(testCase) {
   testCase.set('getDataUrl', getDataUrl);
   const storageLocationsProxy = sinon.stub().resolves(storageLocations);
   testCase.set('storageLocationsProxy', storageLocationsProxy);
+  if (!testCase.get('browserModel')) {
+    testCase.set('browserModel', FilesystemBrowserModel.create({
+      ownerSource: testCase.owner,
+    }));
+  }
 }
 
 async function givenApiSamplesForSharedFile(testCase) {